Move product related components to molecules and organisms
